Use flat arrays as collision grid.
Don't use the sucky array of arrays as collision grid but a simple
linear area containing row after row of data.

The array is by default assumed to be square but can be of of different
width and height if those are defined by .setGrid(array,w,h)

This commit changes the tests to mocha and to use flat grids while
fulfilling the original test specifications.